+++ api/specs/webserver/scripts/main.py
@@ -34,13 +34,12 @@ openapi_modules = [
         "openapi_projects_metadata",
         "openapi_projects_nodes",
         "openapi_projects_ports",
+        "openapi_projects_states",
         "openapi_projects_tags",
-        # "openapi_projects",
         "openapi_publications",
         "openapi_resource_usage",
         "openapi_storage",
         "openapi_tags",
-        # "openapi_tasks",
         "openapi_users",
         "openapi_version_control",
     )
